pyproject.toml
[project]
name = "ytmusic-from-terminal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow==12.0.0",
    "ytmusicapi==1.11.1",
]
